added endpoint /select with parameters & paging

+++ src/main/java/com/exercise/ej2/infrastructure/PersonaListaOutputDTO.java
@@ -3,7 +3,7 @@ package com.exercise.ej2.infrastructure;
 import java.util.List;
 
 public record PersonaListaOutputDTO(
-        int total_items,
+        long total_items,
         List<PersonaOutputDTO> items
 ) {
 }
